+++ test/keypin/di_test.clj
@@ -1,0 +1,24 @@
+;   Copyright (c) Shantanu Kumar. All rights reserved.
+;   The use and distribution terms for this software are covered by the
+;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
+;   which can be found in the file LICENSE at the root of this distribution.
+;   By using this software in any fashion, you are agreeing to be bound by
+;   the terms of this license.
+;   You must not remove this notice, or any other, from this software.
+
+
+(ns keypin.di-test
+  (:require
+    [clojure.test :refer :all]
+    [keypin.di :as di]))
+
+
+(defn foo [a b c] [a b c])
+
+
+(deftest test-partial
+  (let [cc-p (partial foo :first)
+        di-p (di/partial foo :first)]
+    (with-redefs [foo (fn [a b c] {:data [a b c]})]
+      (is (= (cc-p :second :third) [:first :second :third]))
+      (is (= (di-p :second :third) {:data [:first :second :third]})))))