+++ CHANGES.md
@@ -13,7 +13,7 @@
     - Hierarchical config resolution support for all types of config files
   - Support for writing config files
 - Deprecated API
-  - [TODO] Deprecate class `keypin.PropertyFile`
+  - Deprecate class `keypin.PropertyFile`
   - [TODO] Deprecate `keypin.core/read-properties` and `keypin.core/lookup-property`
 - Dependency injection
   - [TODO] Helpers to partially apply annotated arguments in a function