+++ CHANGES.md
@@ -7,8 +7,10 @@
 
 ## 0.4.0 / 2016-June-??
 
+- Project
+  - [TODO] Change project description to encompass configuration handling
 - Config files support
-  - Support for reading config from EDN files
+  - Support for reading config from EDN files ([TODO] tests for EDN)
   - Unified interface for resolving config from various kinds of files (JSON, YAML etc.)
     - Hierarchical config resolution support for all types of config files
   - Support for writing config files
@@ -16,8 +18,9 @@
   - Deprecate class `keypin.PropertyFile`
   - Deprecate `keypin.core/read-properties` and `keypin.core/lookup-property`
 - Dependency injection
-  - Helper fn `keypin.di/partial` to partially apply args to fn var without losing var association
-  - [TODO] Helpers to partially apply annotated arguments in a function
+  - Helper `keypin.di/partial` to partially apply args to fn var without losing var association
+  - Helpers `keypin.di/inject*` and `keypin.di/inject` to partially apply annotated arguments in a function
+  - [TODO] Helpers to partially apply annotated defn arguments in a cascaded fashion
 
 
 ## 0.3.2 / 2016-June-03